+++ lib/src/core/enums/enums.dart
@@ -7,3 +7,6 @@ enum ModifyStrategy { absolute, relative }
 
 /// Defines whether version will be bumped up or down by 1
 enum BumpType { up, down }
+
+/// Filetype being read
+enum FileType { yaml, json, unknown }