+++ lib/src/core/mixins/handle_file_mixin.dart
@@ -1,22 +1,22 @@
+import 'dart:convert';
 import 'dart:io';
 
+import 'package:magical_version_bump/src/core/enums/enums.dart';
+import 'package:magical_version_bump/src/core/extensions/extensions.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:yaml/yaml.dart';
 
 /// This mixin reads and updates the yaml file
 mixin HandleFile {
   /// Read yaml file from path. If:
-  ///   1. [requestPath] is true. The user will be prompted for the path-to-file
-  ///   2. [requestPath] is false. Will assume its the current directory and
-  ///      load using relative path.
+  ///   * `requestPath` is true. The user will be prompted for the path-to-file
+  ///   * `requestPath` is false. Will check `setPath` before checking
+  ///      current directory
   ///
-  /// Returns the path and YAML map
-  ///
-  /// Note: Must provide also yaml file in path incase the file name has been
-  /// changed.
-  Future<({String file, String path, YamlMap yamlMap})> readFile({
+  Future<({String file, FileType type, String path, YamlMap yamlMap})>
+      readFile({
     required Logger logger,
-    bool requestPath = false,
+    required bool requestPath,
     String? setPath,
   }) async {
     var path = ''; // path to file
@@ -40,7 +40,12 @@ mixin HandleFile {
 
     readProgress.complete('Read file');
 
-    return (path: path, file: file, yamlMap: yamlMap);
+    return (
+      path: path,
+      type: path.split('.').last.toLowerCase().fileType,
+      file: file,
+      yamlMap: yamlMap,
+    );
   }
 
   /// Save file changes
@@ -48,8 +53,13 @@ mixin HandleFile {
     required String data,
     required String path,
     required Logger logger,
+    required FileType type,
   }) async {
     final saveProgress = logger.progress('Saving changes');
+
+    if (type == FileType.json) {
+      data = json.encode(_convertToMap(data));
+    }
 
     await File(path).writeAsString(data);