+++ lib/src/core/typedefs/typedefs.dart
@@ -1,0 +1,8 @@
+/// Arguments and their respective values
+typedef ArgsAndValues = Map<String, String>;
+
+/// Reason why an error was thrown:
+///
+/// * Key - title (shown in progress)
+/// * Value - error (logged in console)
+typedef InvalidReason = MapEntry<String, String>;