+++ lib/src/utils/arg_sanitizers/change_arg_sanitizer.dart
@@ -1,0 +1,19 @@
+part of 'arg_sanitizer.dart';
+
+/// Preps args for change command.
+/// Uses default method from parent class to validate.
+final class ChangeArgumentSanitizer extends ArgumentSanitizer {
+  /// Prep change args
+  @override
+  ArgsAndValues prepArgs(List<String> args) {
+    final argsAndValues = <String, String>{};
+
+    for (final argument in args) {
+      final value = argument.split('=');
+
+      argsAndValues.addEntries([MapEntry(value.first, value.last)]);
+    }
+
+    return argsAndValues;
+  }
+}